Only update lockscreen after lock
In 35b9ffbabc318462ef758b10daf2aa70875cb4f4, we quit updating the
lockscreen wallpaper. However, since we can detect when the lockscreen
is active, we note when it is. Next time the system is unlocked, change
the lockscreen. This takes us back to the previous behavior, but with
fewer lockscreen sets and the same number of wallpaper sets as the
present behavior.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>